fixed bug with remove minus nodes visitor
