+++ src/lib/layer1_foundations/expression_parser_sajeeb/Visitors/IRTreeChildLinkArgumentVisitor.cpp
@@ -12,27 +12,27 @@
 
 void IRTreeChildLinkArgumentVisitor::visit(plus_node* op_node) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(minus_node* op_node) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(multiply_node* op_node) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(exponent_node* op_node) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(unary_negate_node* op_node) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(variable_node* num_node) {}
@@ -43,7 +43,7 @@ void IRTreeChildLinkArgumentVisitor::visit(number_node* op_node) {}
 
 void IRTreeChildLinkArgumentVisitor::visit(sentinel_node* op_node) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 
@@ -52,31 +52,31 @@ void IRTreeChildLinkArgumentVisitor::visit(sentinel_node* op_node) {
 void IRTreeChildLinkArgumentVisitor::visit(plus_node* op_node,
                                            list<shared_ptr<irtree_node> >::iterator& link) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(minus_node* op_node,
                                            list<shared_ptr<irtree_node> >::iterator& link) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(multiply_node* op_node,
                                            list<shared_ptr<irtree_node> >::iterator& link) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(exponent_node* op_node,
                                            list<shared_ptr<irtree_node> >::iterator& link) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(unary_negate_node* op_node,
                                            list<shared_ptr<irtree_node> >::iterator& link) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
 
 void IRTreeChildLinkArgumentVisitor::visit(variable_node* num_node,
@@ -91,5 +91,5 @@ void IRTreeChildLinkArgumentVisitor::visit(number_node* op_node,
 void IRTreeChildLinkArgumentVisitor::visit(sentinel_node* op_node,
                                            list<shared_ptr<irtree_node> >::iterator& link) {
     for (auto it=op_node->children.begin(); it != op_node->children.end(); ++it)
-        dispatcher::visit(*it, *this);
+        dispatcher::visit(*it, *this, it);
 }
