+++ src/lib/layer1_foundations/expression_parser_sajeeb/Visitors/IRTreeVisitor.h
@@ -122,40 +122,6 @@ public:
     virtual void visit(sentinel_node* node);
 };
 
-void IRTreeVoidReturnTypeVisitorInterface::visit(plus_node* node) {
-    for (shared_ptr<irtree_node>& child : node->children)
-        dispatcher::visit(child, *this);
-}
-
-void IRTreeVoidReturnTypeVisitorInterface::visit(minus_node* node) {
-    for (shared_ptr<irtree_node>& child : node->children)
-        dispatcher::visit(child, *this);
-}
-
-void IRTreeVoidReturnTypeVisitorInterface::visit(multiply_node* node) {
-    for (shared_ptr<irtree_node>& child : node->children)
-        dispatcher::visit(child, *this);
-}
-
-void IRTreeVoidReturnTypeVisitorInterface::visit(exponent_node* node) {
-    for (shared_ptr<irtree_node>& child : node->children)
-        dispatcher::visit(child, *this);
-}
-
-void IRTreeVoidReturnTypeVisitorInterface::visit(unary_negate_node* node) {
-    for (shared_ptr<irtree_node>& child : node->children)
-        dispatcher::visit(child, *this);
-}
-
-void IRTreeVoidReturnTypeVisitorInterface::visit(sentinel_node* node) {
-    for (shared_ptr<irtree_node>& child : node->children)
-        dispatcher::visit(child, *this);
-}
-
-void IRTreeVoidReturnTypeVisitorInterface::visit(variable_node*) {}
-void IRTreeVoidReturnTypeVisitorInterface::visit(parameter_node*) {}
-void IRTreeVoidReturnTypeVisitorInterface::visit(number_node*) {}
-
 
 class IRTreeVoidReturnTypeConstantVisitorInterface {
 public:
